+++ src/pipx/commands/upgrade.py
@@ -12,7 +12,6 @@ from pipx.venv import Venv, VenvContainer
 
 def upgrade(
     venv_dir: Path,
-    package: str,
     pip_args: List[str],
     verbose: bool,
     *,
@@ -28,6 +27,7 @@ def upgrade(
         )
 
     venv = Venv(venv_dir, verbose=verbose)
+    package = venv.main_package_name
 
     if not venv.package_metadata:
         print(
@@ -37,8 +37,7 @@ def upgrade(
         )
         return 0
 
-    package_metadata = venv.get_package_metadata(package)
-    package = package_metadata.package or package
+    package_metadata = venv.package_metadata[package]
 
     if package_metadata.package_or_url is None:
         raise PipxError(f"Internal Error: package {package} has corrupt pipx metadata.")
@@ -62,15 +61,14 @@ def upgrade(
     )
     # TODO 20191026: upgrade injected packages also (Issue #79)
 
-    package_metadata = venv.get_package_metadata(package)
-    package = package_metadata.package or package
+    package_metadata = venv.package_metadata[package]
+
     display_name = f"{package_metadata.package}{package_metadata.suffix}"
     new_version = package_metadata.package_version
 
     expose_apps_globally(
         constants.LOCAL_BIN_DIR,
         package_metadata.app_paths,
-        package,
         force=force,
         suffix=package_metadata.suffix,
     )
@@ -80,7 +78,6 @@ def upgrade(
             expose_apps_globally(
                 constants.LOCAL_BIN_DIR,
                 app_paths,
-                package,
                 force=force,
                 suffix=package_metadata.suffix,
             )
@@ -103,16 +100,18 @@ def upgrade(
 def upgrade_all(
     venv_container: VenvContainer, verbose: bool, *, skip: List[str], force: bool
 ):
-    packages_upgraded = 0
+    venv_error = False
+    venvs_upgraded = 0
     for venv_dir in venv_container.iter_venv_dirs():
-        package = venv_dir.name
         venv = Venv(venv_dir, verbose=verbose)
-        if package in skip or "--editable" in venv.pipx_metadata.main_package.pip_args:
+        if (
+            venv_dir.name in skip
+            or "--editable" in venv.pipx_metadata.main_package.pip_args
+        ):
             continue
         try:
-            packages_upgraded += upgrade(
+            venvs_upgraded += upgrade(
                 venv_dir,
-                package,
                 venv.pipx_metadata.main_package.pip_args,
                 verbose,
                 upgrading_all=True,
@@ -120,9 +119,15 @@ def upgrade_all(
             )
 
         except Exception:
-            logging.error(f"Error encountered when upgrading {package}")
+            venv_error = True
+            logging.error(f"Error encountered when upgrading {venv_dir.name}")
 
-    if packages_upgraded == 0:
+    if venvs_upgraded == 0:
         print(
             f"Versions did not change after running 'pip upgrade' for each package {sleep}"
+        )
+    if venv_error:
+        raise PipxError(
+            "Some packages encountered errors during upgrade. "
+            "See specific error messages above."
         )