+++ src/pipx/venv.py
@@ -2,7 +2,7 @@ import json
 import logging
 import pkgutil
 from pathlib import Path
-from typing import Dict, Generator, List, NamedTuple, Optional, Set
+from typing import Dict, Generator, List, NamedTuple, Set
 
 from pipx.animate import animate
 from pipx.constants import PIPX_SHARED_PTH
@@ -121,17 +121,14 @@ class Venv:
             ] = self.pipx_metadata.main_package
         return return_dict
 
-    def get_package_metadata(self, package: Optional[str] = None) -> PackageInfo:
-        if package is None or package == self.root.name:
-            return self.pipx_metadata.main_package
-
-        try:
-            return self.package_metadata[package]
-        except KeyError:
-            raise PipxError(
-                f"Package is not installed. Expected to find package {package}, "
-                "but it does not exist."
-            )
+    @property
+    def main_package_name(self) -> str:
+        if self.pipx_metadata.main_package.package is None:
+            # This is OK, because if no metadata, we are pipx < v0.15.0.0 and
+            #   venv_name==main_package_name
+            return self.root.name
+        else:
+            return self.pipx_metadata.main_package.package
 
     def create_venv(self, venv_args: List[str], pip_args: List[str]) -> None:
         with animate("creating virtual environment", self.do_animation):
@@ -313,7 +310,7 @@ class Venv:
             apps_of_dependencies=venv_package_metadata.apps_of_dependencies,
             app_paths_of_dependencies=venv_package_metadata.app_paths_of_dependencies,
             package_version=venv_package_metadata.package_version,
-            suffix=suffix or "",
+            suffix=suffix,
         )
         if is_main_package:
             self.pipx_metadata.main_package = package_info
@@ -368,7 +365,7 @@ class Venv:
             include_dependencies=include_dependencies,
             include_apps=include_apps,
             is_main_package=is_main_package,
-            suffix=suffix or "",
+            suffix=suffix,
         )
 
     def _run_pip(self, cmd: List[str]) -> int: