+++ tests/test_inject.py
@@ -1,11 +1,23 @@
 import pytest  # type: ignore
 
-from helpers import run_pipx_cli
+from helpers import mock_legacy_venv, run_pipx_cli
 
 
-def test_simple(pipx_temp_env, capsys):
+def test_inject_simple(pipx_temp_env, capsys):
     assert not run_pipx_cli(["install", "pycowsay"])
     assert not run_pipx_cli(["inject", "pycowsay", "black"])
+
+
+@pytest.mark.parametrize("metadata_version", [None, "0.1"])
+def test_inject_simple_legacy_venv(pipx_temp_env, capsys, metadata_version):
+    assert not run_pipx_cli(["install", "pycowsay"])
+    mock_legacy_venv("pycowsay", metadata_version=metadata_version)
+    if metadata_version is not None:
+        assert not run_pipx_cli(["inject", "pycowsay", "black"])
+    else:
+        # no metadata in venv should result in PipxError with message
+        assert run_pipx_cli(["inject", "pycowsay", "black"])
+        assert "Please uninstall and install" in capsys.readouterr().err
 
 
 def test_spec(pipx_temp_env, capsys):
@@ -14,7 +26,7 @@ def test_spec(pipx_temp_env, capsys):
 
 
 @pytest.mark.parametrize("with_suffix,", [(False,), (True,)])
-def test_include_apps(pipx_temp_env, capsys, with_suffix):
+def test_inject_include_apps(pipx_temp_env, capsys, with_suffix):
     install_args = []
     suffix = ""